+++ Core/Cursor/CursorManager.cs
@@ -28,9 +28,10 @@ public class CursorManager : MonoBehaviour
     [Header("Touch Mode")]
     [SerializeField] private CursorLockMode touchMode = CursorLockMode.Locked;
 
-    private static Dictionary<string, GameObject> visibleSourcesRequestingCursor = new Dictionary<string, GameObject>();
-
+    private static List<string> visibleSources = new List<string>();
     private static ControlSchemeChangedMessage changeMesage;
+    [Header("Debug")]
+    [SerializeField, EditorReadOnly] private List<string> debugVisibleSources = new List<string>();
     private void Start()
     {
         if (instance != null)
@@ -50,22 +51,28 @@ public class CursorManager : MonoBehaviour
 
     private void Update()
     {
-        //occasionally look for null gameobjects
-        var sourceCount = visibleSourcesRequestingCursor.Count;
+        //TODO: Is there a better solution? Does this matter?
+        //The dictionary will always be very small(under 10 items)
+        //Would this be better/faster as a list?
+        //Does this even matter? This should never happen if handled correctly
 
-        visibleSourcesRequestingCursor = visibleSourcesRequestingCursor
-            .Where(x => x.Value == null)
-            .ToDictionary(t => t.Key, t => t.Value);
-        var newCount = visibleSourcesRequestingCursor.Count;
+        //occasionally look for null gameobjects to prevent bugs where there's
+        //a record, but it's null so it's requesting visible when it shouldn't
+        //var sourceCount = visibleSourcesRequestingCursor.Count;
+        //visibleSourcesRequestingCursor = visibleSourcesRequestingCursor
+        //    .Where(x => x.Value == null)
+        //    .ToDictionary(t => t.Key, t => t.Value);
+        //var newCount = visibleSourcesRequestingCursor.Count;
 
         //there's a change in count, we need to see if we
         //need to update the cursor state
-        if (sourceCount != newCount)
-        {
-            this.UpdateCursorState(ControlSchemeManager.CurrentSchemeType);
-        }
-    }
+        //if (sourceCount != newCount)
+        //{
+        //}
 
+        debugVisibleSources = visibleSources;
+        this.UpdateCursorState(ControlSchemeManager.CurrentSchemeType);
+    }
 
     private void UpdateCursorState(ControlSchemeType newType)
     {
@@ -75,7 +82,14 @@ public class CursorManager : MonoBehaviour
                 Cursor.lockState = defaultMode;
                 break;
             case ControlSchemeType.KBM:
-                Cursor.lockState = kbmMode;
+                if (visibleSources.Count > 0)
+                {
+                    Cursor.lockState = kbmMode_sourcesVisible;
+                }
+                else
+                {
+                    Cursor.lockState = kbmMode;
+                }
                 break;
             case ControlSchemeType.Gamepad:
                 Cursor.lockState = gamepadMode;
@@ -88,19 +102,29 @@ public class CursorManager : MonoBehaviour
         }
     }
 
-    public static void RequestSourceVisible(string name, GameObject go)
+    public static void CursorSourceVisible(string name)
     {
-        if (!visibleSourcesRequestingCursor.ContainsKey(name))
+        //just in case we don't actually have the cursor manager in a scene
+        if (visibleSources == null)
         {
-            visibleSourcesRequestingCursor.Add(name, go);
+            visibleSources = new List<string>();
+        }
+        if (!visibleSources.Contains(name))
+        {
+            visibleSources.Add(name);
             instance.UpdateCursorState(ControlSchemeManager.CurrentSchemeType);
         }
     }
-    public static void UnRequestSourceVisible(string name, GameObject go)
+    public static void CursorSourceHidden(string name)
     {
-        if (visibleSourcesRequestingCursor.ContainsKey(name))
+        //just in case we don't actually have the cursor manager in a scene
+        if (visibleSources == null)
         {
-            visibleSourcesRequestingCursor.Remove(name);
+            visibleSources = new List<string>();
+        }
+        if (visibleSources.Contains(name))
+        {
+            visibleSources.Remove(name);
             instance.UpdateCursorState(ControlSchemeManager.CurrentSchemeType);
         }
     }